talkative_client/pyproject.toml
[tool.poetry]
name = "talkative_client"
version = "0.1.0"
description = "It is example project for study"
authors = ["MaxST <mstolpasov@gmail.com>"]
homepage = "https://github.com/mom1/messager"
repository = "https://github.com/mom1/messager"
include = ["config/**/*"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.1"
sqlalchemy = "^1.3"
sqlalchemy-utils = "^0.34.1"
passlib = "^1.7"
pyqt5 = "^5.13"
dynaconf = "^2.0"
tabulate = "^0.8.3"
pycryptodomex = "^3.8"


[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
